calling worker method to submit job

+++ LibPADIMapNoReduce/Entities/FileSplitMetadata.cs
@@ -13,7 +13,7 @@ namespace PADIMapNoReduce
         int endPosition;
         String clientUrl;//workers need this to connect to client
 
-      public  FileSplitMetadata(int splitId, int startPosition, int endPosition, String clientUrl)
+      public  FileSplitMetadata(int splitId, long startPosition, long endPosition, String clientUrl)
         {
             this.splitId = splitId;
             this.startPosition = startPosition;
